update users module in admin dashboard

+++ database/migrations/2023_06_13_085649_create_wishlists_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->increments('id');
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
-            $table->unsignedBigInteger('unit_id');
+            $table->unsignedInteger('unit_id');
             $table->foreign('unit_id')->references('id')->on('units');
             $table->timestamps();
         });
